refactor: encapsulated vnode mapping struct (#7505)
This PR introduces an encapsulated structure for `VnodeMapping`, which uses the compressed format which can be efficiently converted into/from the protobuf representation, and can be indexed with logarithmic complexity. See `src/common/src/hash/consistent_hash/mapping.rs` for more details.

This PR also distinguishes `pb::ParallelUnitMapping` from `pb::FragmentParallelUnitMapping`. The notification requires a single message which contains the key and value, so we previously put the fragment ID into the `pb::ParallelUnitMapping`, while the `pb::Fragment.vnode_mapping` also contains the fragment ID which makes no sense.

Approved-By: xx01cyx
Approved-By: yezizp2012

+++ src/common/src/util/compress.rs
@@ -17,7 +17,7 @@ use itertools::Itertools;
 /// This function compresses sequential repeated data in a vector. The compression result contains
 /// two vectors, one for the last indices of sequential repeated elements, and another for the
 /// repeated data. For example, [14, 14, 14, 27, 27] will be compressed to [2, 4], [14, 27].
-pub fn compress_data<T>(original_data: &[T]) -> (Vec<u64>, Vec<T>)
+pub fn compress_data<T>(original_data: &[T]) -> (Vec<u32>, Vec<T>)
 where
     T: PartialEq + Copy,
 {
@@ -26,13 +26,13 @@ where
 
     for i in 1..original_data.len() {
         if original_data[i - 1] != original_data[i] {
-            original_indices.push(i as u64 - 1);
+            original_indices.push(i as u32 - 1);
             data.push(original_data[i - 1]);
         }
     }
 
     if let Some(&last) = original_data.last() {
-        original_indices.push(original_data.len() as u64 - 1);
+        original_indices.push(original_data.len() as u32 - 1);
         data.push(last);
     }
 
@@ -40,7 +40,7 @@ where
 }
 
 /// Works in a reversed way as `compress_data`.
-pub fn decompress_data<T>(original_indices: &[u64], data: &[T]) -> Vec<T>
+pub fn decompress_data<T>(original_indices: &[u32], data: &[T]) -> Vec<T>
 where
     T: Copy,
 {
@@ -65,7 +65,7 @@ mod tests {
         // Simple
         let original_data = [3u32, 3, 3, 3, 3, 4, 4, 5, 5, 6, 7, 8, 8, 8, 9];
         let (compressed_original_indices, compressed_data) = compress_data(&original_data);
-        let expect_original_indices = Vec::from([4u64, 6, 8, 9, 10, 13, 14]);
+        let expect_original_indices = Vec::from([4u32, 6, 8, 9, 10, 13, 14]);
         let expect_data = Vec::from([3u32, 4, 5, 6, 7, 8, 9]);
         assert_eq!(compressed_original_indices, expect_original_indices);
         assert_eq!(compressed_data, expect_data);
@@ -81,7 +81,7 @@ mod tests {
         long_original_data[0] = 5;
         long_original_data[2046] = 5;
         let (compressed_original_indices, compressed_data) = compress_data(&long_original_data);
-        let expect_original_indices = Vec::from([0u64, 511, 1023, 1535, 2045, 2046, 2047]);
+        let expect_original_indices = Vec::from([0u32, 511, 1023, 1535, 2045, 2046, 2047]);
         let expect_data = Vec::from([5u32, 1, 2, 3, 4, 5, 4]);
         assert_eq!(compressed_original_indices, expect_original_indices);
         assert_eq!(compressed_data, expect_data);
